Fix Safari rendering issue with CanvasLand and cropping
CanvasLand renders the field of play to a background canvas, and then copies the part that is on-screen onto the canvas in the document. On Safari, this encountered a bug whereby if the part of the image to be copied went over the bounds of the field of play (e.g. the target canvas was bigger, or it had been moved so it was copying from (-1,-1)) then nothing would be copied.

The fix manually calculates the bounds of the crop area to avoid going over the edge of the field of play.